PLATFORM-3624: move redirect class to the extension code
