+++ extensions/wikia/RobotsTxt/RobotsTxt.setup.php
@@ -3,6 +3,7 @@
 // Autoload
 $wgAutoloadClasses['Wikia\RobotsTxt\RobotsTxt'] =  __DIR__ . '/classes/RobotsTxt.class.php';
 $wgAutoloadClasses['Wikia\RobotsTxt\PathBuilder'] =  __DIR__ . '/classes/PathBuilder.class.php';
+$wgAutoloadClasses['Wikia\RobotsTxt\RobotsRedirect'] =  __DIR__ . '/classes/RobotsRedirect.class.php';
 $wgAutoloadClasses['Wikia\RobotsTxt\WikiaRobots'] =  __DIR__ . '/classes/WikiaRobots.class.php';
 $wgAutoloadClasses['WikiaRobotsController'] =  __DIR__ . '/WikiaRobotsController.class.php';
